# Attempt to speed-up body cells display by reusing Y open axis facts p…