--- a/arelle/ViewWinRenderedGrid.py
+++ b/arelle/ViewWinRenderedGrid.py
@@ -261,7 +261,11 @@ class ViewRenderedGrid(ViewWinTkTable.ViewTkTable):
                 if fp is not None:
                     fp.clear()
             self.factPrototypes = []
+            
+            startedAt2 = time.time()
             self.bodyCells(self.dataFirstRow, yTopStructuralNode, xStructuralNodes, self.zAspectStructuralNodes, self.yAxisChildrenFirst.get())
+            print("bodyCells {:.2f}secs ".format(time.time() - startedAt2) + self.roledefinition)
+            
             self.table.clearModificationStatus()
             self.table.disableUnusedCells()
             self.table.resizeTableCells()
@@ -599,6 +603,13 @@ class ViewRenderedGrid(ViewWinTkTable.ViewTkTable):
                     # data for columns of row
                     #print ("row " + str(row) + "yNode " + yStructuralNode.definitionNode.objectId() )
                     ignoreDimValidity = self.ignoreDimValidity.get()
+                    
+                    # Reuse already computed facts partition in case of open Y axis
+                    if True and hasattr(yStructuralNode, "factsPartition"):
+                        factsPartition = yStructuralNode.factsPartition
+                    else:
+                        factsPartition = None
+                    
                     for i, xStructuralNode in enumerate(xStructuralNodes):
                         isEntryPrototype = isYEntryPrototype or xStructuralNode.isEntryPrototype(default=False)
                         xAspectStructuralNodes = defaultdict(set)
@@ -632,6 +643,10 @@ class ViewRenderedGrid(ViewWinTkTable.ViewTkTable):
                         justify = None
                         fp = FactPrototype(self, cellAspectValues)
                         if conceptNotAbstract:
+                            if factsPartition is not None:
+                                # pick from this set of facts
+                                facts = factsPartition    
+                            else:     
                             # reduce set of matchable facts to those with pri item qname and have dimension aspects
                             facts = self.modelXbrl.factsByQname[priItemQname] if priItemQname else self.modelXbrl.factsInInstance
                             if self.hasTableFilters:
@@ -664,6 +679,9 @@ class ViewRenderedGrid(ViewWinTkTable.ViewTkTable):
                                     else:
                                         value = fact.effectiveValue
                                     objectId = fact.objectId()
+                                    # we can now remove that fact if we picked up from the computed partition entry
+                                    if factsPartition is not None:
+                                        factsPartition.remove(fact)
                                     justify = XbrlTable.TG_RIGHT_JUSTIFIED if fact.isNumeric else XbrlTable.TG_LEFT_JUSTIFIED
                                     break
                         if (conceptNotAbstract and