…artition info computed when resolvin table structures

For 300 items : 50sec -> 7sec
For 1500 items : 1210sec -> 40 sec